Time: 3 ms (91.90%) | Memory: 35 MB (82.06%) - LeetSync

+++ 2090-number-of-ways-to-arrive-at-destination/number-of-ways-to-arrive-at-destination.cpp
@@ -12,7 +12,7 @@ public:
 
         // Use long long for distance and path count to avoid overflow
         vector<long long> dst(n, LLONG_MAX);
-        vector<long long> paths(n, 0);
+        vector<int> paths(n, 0);
 
         // Initializing the source node
         dst[0] = 0;
